+++ frontend/pages/produtos/[id].tsx
@@ -1,19 +1,20 @@
-// frontend/pages/produtos/[id].tsx - edição de produto
+// frontend/pages/produtos/[id].tsx
 import React, { useState, useEffect } from 'react';
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import { Card } from '@/components/ui/Card';
 import { Input } from '@/components/ui/Input';
+import { MaskedInput } from '@/components/ui/MaskedInput';
 import { Select } from '@/components/ui/Select';
 import { RadioGroup } from '@/components/ui/RadioGroup';
 import { Button } from '@/components/ui/Button';
 import { Tabs } from '@/components/ui/Tabs';
 import { useToast } from '@/components/ui/ToastContext';
 import { useRouter } from 'next/router';
-import { PageLoader } from '@/components/ui/PageLoader';
 import api from '@/lib/api';
+import { PageLoader } from '@/components/ui/PageLoader';
+import { formatCPFOrCNPJ, formatCEP } from '@/lib/validation';
 import { Trash2, BrainCog } from 'lucide-react';
 import { useOperadorEstrangeiro, OperadorEstrangeiro } from '@/hooks/useOperadorEstrangeiro';
-import { formatCPFOrCNPJ, formatCEP } from '@/lib/validation';
 import { OperadorEstrangeiroSelector } from '@/components/operadores-estrangeriros/OperadorEstrangeiroSelector';
 
 interface AtributoEstrutura {
@@ -29,7 +30,8 @@ interface AtributoEstrutura {
   subAtributos?: AtributoEstrutura[];
 }
 
-export default function EditarProdutoPage() {
+export default function ProdutoPage() {
+  const [catalogoId, setCatalogoId] = useState('');
   const [catalogoNome, setCatalogoNome] = useState('');
   const [codigo, setCodigo] = useState('');
   const [codigosInternos, setCodigosInternos] = useState<string[]>([]);
@@ -37,21 +39,35 @@ export default function EditarProdutoPage() {
   const [operadores, setOperadores] = useState<Array<{ paisCodigo: string; conhecido: string; operador?: OperadorEstrangeiro | null }>>([]);
   const [novoOperador, setNovoOperador] = useState<{ paisCodigo: string; conhecido: string; operador?: OperadorEstrangeiro | null }>({ paisCodigo: '', conhecido: 'nao', operador: undefined });
   const [selectorOpen, setSelectorOpen] = useState(false);
-  const { getPaisOptions, getPaisNome } = useOperadorEstrangeiro();
+  const { getPaisOptions, buscarOperadorPorId, getPaisNome } = useOperadorEstrangeiro();
+  const [catalogos, setCatalogos] = useState<Array<{ id: number; nome: string; cpf_cnpj: string | null }>>([]);
   const [ncm, setNcm] = useState('');
   const [ncmDescricao, setNcmDescricao] = useState('');
+  const [unidadeMedida, setUnidadeMedida] = useState('');
   const [modalidade, setModalidade] = useState('IMPORTACAO');
   const [denominacao, setDenominacao] = useState('');
   const [descricao, setDescricao] = useState('');
   const [estrutura, setEstrutura] = useState<AtributoEstrutura[]>([]);
   const [valores, setValores] = useState<Record<string, string>>({});
-  const [loading, setLoading] = useState(true);
-  const [activeTab, setActiveTab] = useState('fixos');
+  const [loadingEstrutura, setLoadingEstrutura] = useState(false);
+  const [estruturaCarregada, setEstruturaCarregada] = useState(false);
+  const [activeTab, setActiveTab] = useState('informacoes');
+  const [loading, setLoading] = useState(false);
   const { addToast } = useToast();
   const router = useRouter();
   const { id } = router.query;
+  const isNew = !id || id === 'novo';
 
   // Texto do cabeçalho removido conforme novo layout
+
+  useEffect(() => {
+    if (!isNew) return;
+    async function carregarCatalogos() {
+      const res = await api.get('/catalogos');
+      setCatalogos(res.data);
+    }
+    carregarCatalogos();
+  }, [isNew]);
 
   const mapaEstrutura = React.useMemo(() => {
     const map = new Map<string, AtributoEstrutura>();
@@ -95,13 +111,54 @@ export default function EditarProdutoPage() {
     return resultado;
   }
 
-  async function carregarEstrutura() {
-    if (ncm.length < 8) return;
-    const response = await api.get(`/siscomex/atributos/ncm/${ncm}?modalidade=${modalidade}`);
-    const dados: AtributoEstrutura[] = response.data.dados || [];
-    setNcmDescricao(response.data.descricaoNcm || '');
-    setEstrutura(ordenarAtributos(dados));
+  async function carregarEstrutura(codigo?: string) {
+    const ncmCodigo = codigo || ncm;
+    if (ncmCodigo.length < 8) return;
+    setLoadingEstrutura(true);
+    try {
+      const response = await api.get(
+        `/siscomex/atributos/ncm/${ncmCodigo}?modalidade=${modalidade}`
+      );
+
+      if (!response.data.descricaoNcm) {
+        addToast('NCM não encontrada', 'error');
+        setEstruturaCarregada(false);
+        setNcmDescricao('');
+        setUnidadeMedida('');
+        setEstrutura([]);
+        return;
+      }
+
+      const dados: AtributoEstrutura[] = response.data.dados || [];
+      setNcmDescricao(response.data.descricaoNcm);
+      setUnidadeMedida(response.data.unidadeMedida || '');
+      setEstrutura(ordenarAtributos(dados));
+      setEstruturaCarregada(true);
+    } catch (error) {
+      console.error('Erro ao carregar atributos:', error);
+      addToast('Erro ao carregar atributos', 'error');
+      setEstruturaCarregada(false);
+    } finally {
+      setLoadingEstrutura(false);
+    }
   }
+
+  function handleNcmChange(valor: string) {
+    setNcm(valor);
+    if (valor.length === 8) {
+      carregarEstrutura(valor);
+    } else {
+      setNcmDescricao('');
+      setUnidadeMedida('');
+      setEstruturaCarregada(false);
+    }
+  }
+
+  useEffect(() => {
+    if (ncm.length === 8) {
+      carregarEstrutura(ncm);
+    }
+  }, [modalidade]);
 
   function handleValor(codigo: string, valor: string) {
     setValores(prev => ({ ...prev, [codigo]: valor }));
@@ -130,7 +187,7 @@ export default function EditarProdutoPage() {
 
   function formatarEndereco(op?: OperadorEstrangeiro | null) {
     if (!op) return '';
-    const partes: string[] = [];
+    const partes = [] as string[];
     if (op.logradouro) partes.push(op.logradouro);
     if (op.cidade) partes.push(op.cidade);
     if (op.subdivisao) partes.push(op.subdivisao.nome);
@@ -169,6 +226,8 @@ export default function EditarProdutoPage() {
     return ok;
   }
 
+  // A visibilidade dos atributos condicionados é verificada a cada alteração
+  // pois o componente re-renderiza sempre que 'valores' é atualizado.
   function condicaoAtendida(attr: AtributoEstrutura): boolean {
     if (!attr.parentCodigo) return true;
 
@@ -203,7 +262,8 @@ export default function EditarProdutoPage() {
             label={attr.nome}
             required={attr.obrigatorio}
             options={
-              attr.dominio?.map(d => ({ value: d.codigo, label: d.descricao })) || []
+              attr.dominio?.map(d => ({ value: d.codigo, label: d.descricao })) ||
+              []
             }
             placeholder="Selecione..."
             value={value}
@@ -271,6 +331,7 @@ export default function EditarProdutoPage() {
 
   async function carregarProduto(produtoId: string) {
     try {
+      setLoading(true);
       const response = await api.get(`/produtos/${produtoId}`);
       const dados = response.data;
       setCodigo(dados.codigo);
@@ -285,6 +346,7 @@ export default function EditarProdutoPage() {
       setDenominacao(dados.denominacao || '');
       setDescricao(dados.descricao || '');
       setCatalogoNome(dados.catalogo?.nome || '');
+      setCatalogoId(String(dados.catalogo?.id || ''));
       setNcm(dados.ncmCodigo);
       setModalidade(dados.modalidade);
       try {
@@ -292,11 +354,14 @@ export default function EditarProdutoPage() {
           `/siscomex/atributos/ncm/${dados.ncmCodigo}?modalidade=${dados.modalidade}`
         );
         setNcmDescricao(resp.data.descricaoNcm || '');
+        setUnidadeMedida(resp.data.unidadeMedida || '');
       } catch (e) {
         setNcmDescricao('');
+        setUnidadeMedida('');
       }
       const estr = dados.atributos?.[0]?.estruturaSnapshotJson || [];
       setEstrutura(ordenarAtributos(estr));
+      setEstruturaCarregada(true);
       setValores(dados.atributos?.[0]?.valoresJson || {});
     } catch (error) {
       console.error('Erro ao carregar produto:', error);
@@ -308,15 +373,20 @@ export default function EditarProdutoPage() {
   }
 
   useEffect(() => {
-    if (router.isReady && typeof id === 'string') {
-      carregarProduto(id);
+    if (!router.isReady) return;
+    if (!isNew && typeof id === 'string') {
+      carregarProduto(id as string);
     }
-  }, [router.isReady, id]);
+  }, [router.isReady, id, isNew]);
 
   async function salvar() {
     try {
-      await api.put(`/produtos/${id}`, {
+      const url = isNew ? '/produtos' : `/produtos/${id}`;
+      const metodo = isNew ? api.post : api.put;
+      await metodo(url, {
+        ncmCodigo: ncm,
         modalidade,
+        catalogoId: catalogoId ? Number(catalogoId) : undefined,
         denominacao,
         descricao,
         valoresAtributos: valores,
@@ -327,7 +397,7 @@ export default function EditarProdutoPage() {
           operadorEstrangeiroId: o.operador?.id
         }))
       });
-      addToast('Produto atualizado com sucesso!', 'success');
+      addToast(isNew ? 'Produto salvo com sucesso!' : 'Produto atualizado com sucesso!', 'success');
       router.push('/produtos');
     } catch (error: any) {
       handleApiError(error);
@@ -342,7 +412,7 @@ export default function EditarProdutoPage() {
           return `${nome}: ${d.message}`;
         })
         .join('; ');
-      addToast(`Erro de validação: ${details}`, 'error');
+      addToast(`Erro de valida\u00e7\u00e3o: ${details}`, 'error');
     } else if (error.response?.data?.error) {
       addToast(error.response.data.error, 'error');
     } else {
@@ -350,20 +420,20 @@ export default function EditarProdutoPage() {
     }
   }
 
-  if (loading) {
-    return (
-      <DashboardLayout title="Editar Produto">
-        <PageLoader message="Carregando produto..." />
-      </DashboardLayout>
-    );
-  }
-
   return (
-    <DashboardLayout title="Editar Produto">
-      <Card className="mb-6" headerTitle="Seleção do Catálogo">
-        <div className="grid grid-cols-3 gap-4">
-          <Input label="Catálogo" value={catalogoNome} disabled />
-          <Input label="NCM" value={ncm} disabled />
+    <DashboardLayout title={isNew ? 'Novo Produto' : 'Editar Produto'}>
+      <Card className="mb-6" headerTitle="Catálogo / NCM">
+        <div className="grid grid-cols-4 gap-4">
+          {isNew ? (
+            <Select
+              label="Catálogo"
+              options={catalogos.map(c => ({ value: String(c.id), label: `${c.nome} - ${formatCPFOrCNPJ(c.cpf_cnpj)}` }))}
+              value={catalogoId}
+              onChange={e => setCatalogoId(e.target.value)}
+            />
+          ) : (
+            <Input label="Catálogo" value={catalogoNome} disabled />
+          )}
           <Select
             label="Modalidade"
             options={[
@@ -373,178 +443,228 @@ export default function EditarProdutoPage() {
             value={modalidade}
             onChange={e => setModalidade(e.target.value)}
           />
-          <div className="flex items-end">
-            <Button type="button" onClick={carregarEstrutura}>Carregar Estrutura</Button>
-          </div>
+          {estruturaCarregada && !loadingEstrutura && (
+              <div className="grid grid-cols-1 gap-4">
+                <Input label="Código" value={codigo || '-'} disabled />
+              </div>
+            )}
         </div>
+
+        {catalogoId && (
+          <>
+            <div className="grid grid-cols-5 gap-4 mt-4">
+              {isNew ? (
+                <MaskedInput
+                  label="NCM"
+                  mask="ncm"
+                  value={ncm}
+                  onChange={val => handleNcmChange(val)}
+                  className="col-span-1"
+                />
+              ) : (
+                <Input label="NCM" value={ncm} disabled className="col-span-1" />
+              )}
+              <Input label="Descrição NCM" value={ncmDescricao} disabled className="col-span-3" />
+              <Input label="Unidade de Medida" value={unidadeMedida} disabled className="col-span-1" />
+            </div>
+          </>
+        )}
       </Card>
 
-      <Card className="mb-6">
-        <Tabs
-          activeId={activeTab}
-          onChange={setActiveTab}
-          tabs={[
-            {
-              id: 'fixos',
-              label: 'Dados Fixos',
-              content: (
-                <div className="grid grid-cols-3 gap-4 text-sm">
-                  <Input label="Código" value={codigo || '-'} disabled />
-                  <Input
-                    label="Nome do Produto"
-                    className="col-span-3"
-                    value={denominacao}
-                    onChange={e => setDenominacao(e.target.value)}
-                  />
-                  <div className="col-span-3 mb-4">
-                    <label htmlFor="descricao" className="block text-sm font-medium mb-1 text-gray-300">
-                      Descrição
-                    </label>
-                    <textarea
-                      id="descricao"
-                      className="w-full px-2 py-1 text-sm bg-[#1e2126] border border-gray-700 text-white rounded-md focus:outline-none focus:ring focus:border-blue-500"
-                      placeholder="Descrição do Produto"
-                      rows={4}
-                      value={descricao}
-                      onChange={e => setDescricao(e.target.value)}
-                    />
-                    <div className="mt-2">
-                      <Button type="button" size="sm" onClick={() => setActiveTab('dinamicos')}>
-                        <BrainCog size={16} className="inline mr-2" /> Preencher Atributos
-                      </Button>
-                    </div>
-                  </div>
+      {catalogoId && (
+        <>
 
-                  <Card
-                    headerTitle="Códigos Internos"
-                    headerClassName="bg-[#1a1f2b] px-4 py-2"
-                    className="col-span-3 mt-2"
-                  >
-                    <div className="flex gap-2 mb-4">
-                      <Input
-                        value={novoCodigoInterno}
-                        onChange={e => setNovoCodigoInterno(e.target.value)}
-                        className="mb-0 flex-1"
-                      />
-                      <Button type="button" onClick={adicionarCodigoInterno}>+ Incluir</Button>
-                    </div>
-                    {codigosInternos.length > 0 && (
-                      <div className="overflow-x-auto">
-                        <table className="w-full text-xs text-left">
-                          <thead className="text-gray-400 bg-[#0f1419] uppercase">
-                            <tr>
-                              <th className="w-16 px-4 py-2 text-center">Ações</th>
-                              <th className="px-4 py-2">Código</th>
-                            </tr>
-                          </thead>
-                          <tbody>
-                            {codigosInternos.map((c, i) => (
-                              <tr key={i} className="border-b border-gray-700">
-                                <td className="px-4 py-1 text-center">
-                                  <button
-                                    className="p-1 text-gray-300 hover:text-red-500 transition-colors"
-                                    onClick={() => removerCodigoInterno(i)}
-                                  >
-                                    <Trash2 size={16} />
-                                  </button>
-                                </td>
-                                <td className="px-4 py-1">{c}</td>
-                              </tr>
-                            ))}
-                          </tbody>
-                        </table>
-                      </div>
-                    )}
-                  </Card>
+          {loadingEstrutura && (
+            <Card className="mb-6">
+              <PageLoader message="Carregando dados do produto..." />
+            </Card>
+          )}
 
-                  <Card
-                    headerTitle="Operadores Estrangeiros"
-                    headerClassName="bg-[#1a1f2b] px-4 py-2"
-                    className="col-span-3 mt-4"
-                  >
-                    <div className="grid grid-cols-3 gap-4 mb-4">
-                      <Select
-                        label="País"
-                        options={getPaisOptions()}
-                        value={novoOperador.paisCodigo}
-                        onChange={e => setNovoOperador(prev => ({ ...prev, paisCodigo: e.target.value }))}
-                        className="mb-0"
-                      />
-                      <RadioGroup
-                        label="Conhecido?"
-                        options={[{ value: 'nao', label: 'Não' }, { value: 'sim', label: 'Sim' }]}
-                        value={novoOperador.conhecido}
-                        onChange={v => setNovoOperador(prev => ({ ...prev, conhecido: v }))}
-                        className="mb-0"
-                      />
-                      {novoOperador.conhecido === 'sim' && (
-                        <div className="flex items-end gap-2">
-                          <Input label="Operador" value={novoOperador.operador?.nome || ''} readOnly className="flex-1" />
-                          <Button type="button" onClick={() => setSelectorOpen(true)}>Buscar</Button>
+          {estruturaCarregada && !loadingEstrutura && (
+            <>
+              <Card className="mb-6">
+                <Tabs
+                  activeId={activeTab}
+                  onChange={setActiveTab}
+                  tabs={[
+                    {
+                      id: 'informacoes',
+                      label: 'Informações do Produto',
+                      content: (
+                        <div className="grid grid-cols-3 gap-4 text-sm">
+                          <Input
+                            label="Nome do Produto"
+                            className="col-span-3"
+                            value={denominacao}
+                            onChange={e => setDenominacao(e.target.value)}
+                          />
+
+                          <div className="col-span-3 mb-4">
+                            <label htmlFor="descricao" className="block text-sm font-medium mb-1 text-gray-300">
+                              Descrição
+                            </label>
+                            <textarea
+                              id="descricao"
+                              className="w-full px-2 py-1 text-sm bg-[#1e2126] border border-gray-700 text-white rounded-md focus:outline-none focus:ring focus:border-blue-500"
+                              placeholder="Descrição do Produto"
+                              rows={4}
+                              value={descricao}
+                              onChange={e => setDescricao(e.target.value)}
+                            />
+                            <div className="mt-2">
+                              <Button type="button" size="sm" onClick={() => setActiveTab('dinamicos')}>
+                                <BrainCog size={16} className="inline mr-2" /> Preencher Atributos
+                              </Button>
+                            </div>
+                          </div>
+
+                          <div className="col-span-3">
+                          <Card
+                            headerTitle="Códigos Internos"
+                            headerClassName="bg-[#1a1f2b] px-4 py-2"
+                          >
+                              <div className="flex gap-2 mb-4">
+                                <Input
+                                  value={novoCodigoInterno}
+                                  onChange={e => setNovoCodigoInterno(e.target.value)}
+                                  className="mb-0 w-1/2"
+                                />
+                                <Button type="button" onClick={adicionarCodigoInterno}>+ Incluir</Button>
+                              </div>
+                              {codigosInternos.length > 0 && (
+                                <div className="overflow-x-auto">
+                                  <table className="w-full text-xs text-left">
+                                    <thead className="text-gray-400 bg-[#0f1419] uppercase">
+                                      <tr>
+                                        <th className="w-16 px-4 py-2 text-center">Ações</th>
+                                        <th className="px-4 py-2">Código</th>
+                                      </tr>
+                                    </thead>
+                                    <tbody>
+                                      {codigosInternos.map((c, i) => (
+                                        <tr key={i} className="border-b border-gray-700">
+                                          <td className="px-4 py-1 text-center">
+                                            <button
+                                              className="p-1 text-gray-300 hover:text-red-500 transition-colors"
+                                              onClick={() => removerCodigoInterno(i)}
+                                            >
+                                              <Trash2 size={16} />
+                                            </button>
+                                          </td>
+                                          <td className="px-4 py-1">{c}</td>
+                                        </tr>
+                                      ))}
+                                    </tbody>
+                                  </table>
+                                </div>
+                              )}
+                          </Card>
+                          <Card
+                            headerTitle="Operadores Estrangeiros"
+                            headerClassName="bg-[#1a1f2b] px-4 py-2"
+                            className="col-span-3 mt-4"
+                          >
+                            <div className="grid grid-cols-3 gap-4 mb-4">
+                              <Select
+                                label="País"
+                                options={getPaisOptions()}
+                                value={novoOperador.paisCodigo}
+                                onChange={e => setNovoOperador(prev => ({ ...prev, paisCodigo: e.target.value }))}
+                                className="mb-0"
+                              />
+                              <RadioGroup
+                                label="Conhecido?"
+                                options={[{ value: 'nao', label: 'Não' }, { value: 'sim', label: 'Sim' }]}
+                                value={novoOperador.conhecido}
+                                onChange={v => setNovoOperador(prev => ({ ...prev, conhecido: v }))}
+                                className="mb-0"
+                              />
+                              {novoOperador.conhecido === 'sim' && (
+                                <div className="flex items-end gap-2">
+                                  <Input
+                                    label="Operador"
+                                    value={novoOperador.operador?.nome || ''}
+                                    readOnly
+                                    className="flex-1 mb-0"
+                                  />
+                                  <Button type="button" onClick={() => setSelectorOpen(true)}>
+                                    Buscar
+                                  </Button>
+                                </div>
+                              )}
+                            </div>
+                            <Button type="button" onClick={adicionarOperador}>Vincular Operador</Button>
+
+                            {operadores.length > 0 && (
+                              <div className="overflow-x-auto mt-4">
+                                <table className="w-full text-xs text-left">
+                                  <thead className="text-gray-400 bg-[#0f1419] uppercase">
+                                    <tr>
+                                      <th className="w-16 px-4 py-2 text-center">Ações</th>
+                                      <th className="px-4 py-2">País de Origem</th>
+                                      <th className="px-4 py-2">Conhecido</th>
+                                      <th className="px-4 py-2">CPF/CNPJ/TIN</th>
+                                      <th className="px-4 py-2">Código</th>
+                                      <th className="px-4 py-2">Código Interno</th>
+                                      <th className="px-4 py-2">Nome</th>
+                                      <th className="px-4 py-2">Endereço</th>
+                                    </tr>
+                                  </thead>
+                                  <tbody>
+                                    {operadores.map((op, i) => (
+                                      <tr key={i} className="border-b border-gray-700">
+                                        <td className="px-4 py-1 text-center">
+                                          <button className="p-1 text-gray-300 hover:text-red-500 transition-colors" onClick={() => removerOperador(i)}>
+                                            <Trash2 size={16} />
+                                          </button>
+                                        </td>
+                                        <td className="px-4 py-1">{op.operador?.pais?.nome || getPaisNome(op.paisCodigo)}</td>
+                                        <td className="px-4 py-1">{op.conhecido === 'sim' ? 'Sim' : 'Não'}</td>
+                                        <td className="px-4 py-1">
+                                          {op.conhecido === 'sim' ? (op.operador?.tin || formatCPFOrCNPJ(op.operador?.cnpjRaizResponsavel)) : ''}
+                                        </td>
+                                        <td className="px-4 py-1">{op.conhecido === 'sim' ? op.operador?.codigo || '' : ''}</td>
+                                        <td className="px-4 py-1">{op.conhecido === 'sim' ? op.operador?.codigoInterno || '' : ''}</td>
+                                        <td className="px-4 py-1">{op.conhecido === 'sim' ? op.operador?.nome || '' : ''}</td>
+                                        <td className="px-4 py-1">{op.conhecido === 'sim' ? formatarEndereco(op.operador) : ''}</td>
+                                      </tr>
+                                    ))}
+                                  </tbody>
+                                </table>
+                              </div>
+                            )}
+                          </Card>
                         </div>
-                      )}
-                    </div>
-                    <Button type="button" onClick={adicionarOperador}>Vincular Operador</Button>
-
-                    {operadores.length > 0 && (
-                      <div className="overflow-x-auto mt-4">
-                        <table className="w-full text-xs text-left">
-                          <thead className="text-gray-400 bg-[#0f1419] uppercase">
-                            <tr>
-                              <th className="w-16 px-4 py-2 text-center">Ações</th>
-                              <th className="px-4 py-2">País de Origem</th>
-                              <th className="px-4 py-2">Conhecido</th>
-                              <th className="px-4 py-2">CPF/CNPJ/TIN</th>
-                              <th className="px-4 py-2">Código</th>
-                              <th className="px-4 py-2">Código Interno</th>
-                              <th className="px-4 py-2">Nome</th>
-                              <th className="px-4 py-2">Endereço</th>
-                            </tr>
-                          </thead>
-                          <tbody>
-                            {operadores.map((op, i) => (
-                              <tr key={i} className="border-b border-gray-700">
-                                <td className="px-4 py-1 text-center">
-                                  <button className="p-1 text-gray-300 hover:text-red-500 transition-colors" onClick={() => removerOperador(i)}>
-                                    <Trash2 size={16} />
-                                  </button>
-                                </td>
-                                <td className="px-4 py-1">{op.operador?.pais?.nome || getPaisNome(op.paisCodigo)}</td>
-                                <td className="px-4 py-1">{op.conhecido === 'sim' ? 'Sim' : 'Não'}</td>
-                                <td className="px-4 py-1">
-                                  {op.conhecido === 'sim' ? (op.operador?.tin || formatCPFOrCNPJ(op.operador?.cnpjRaizResponsavel)) : ''}
-                                </td>
-                                <td className="px-4 py-1">{op.conhecido === 'sim' ? op.operador?.codigo || '' : ''}</td>
-                                <td className="px-4 py-1">{op.conhecido === 'sim' ? op.operador?.codigoInterno || '' : ''}</td>
-                                <td className="px-4 py-1">{op.conhecido === 'sim' ? op.operador?.nome || '' : ''}</td>
-                                <td className="px-4 py-1">{op.conhecido === 'sim' ? formatarEndereco(op.operador) : ''}</td>
-                              </tr>
-                            ))}
-                          </tbody>
-                        </table>
                       </div>
-                    )}
-                  </Card>
-                </div>
-              )
-            },
-            {
-              id: 'dinamicos',
-              label: 'Atributos Dinâmicos',
-              content: (
-                <div className="grid grid-cols-3 gap-4 text-sm">
-                  {estrutura.map(attr => renderCampo(attr))}
-                </div>
-              )
-            }
-          ]}
-        />
-      </Card>
+                    )
+                  },
+                    {
+                      id: 'dinamicos',
+                      label: 'Atributos Dinâmicos',
+                      content: (
+                        <div className="grid grid-cols-3 gap-4 text-sm">
+                          {estrutura.map(attr => renderCampo(attr))}
+                        </div>
+                      )
+                    }
+                  ]}
+                />
+              </Card>
 
-      <div className="flex justify-end gap-3">
-        <Button type="button" variant="outline" onClick={() => router.push('/produtos')}>Cancelar</Button>
-        <Button type="button" onClick={salvar}>Salvar Produto</Button>
-      </div>
+              <div className="flex justify-end gap-3">
+                <Button
+                  type="button"
+                  variant="outline"
+                  onClick={() => router.push('/produtos')}
+                >
+                  Cancelar
+                </Button>
+                <Button type="button" onClick={salvar}>Salvar Produto</Button>
+              </div>
+            </>
+          )}
+        </>
+      )}
       {selectorOpen && (
         <OperadorEstrangeiroSelector
           onSelect={op => { setNovoOperador(prev => ({ ...prev, operador: op })); setSelectorOpen(false); }}
